+++ titiler/pgstac/main.py
@@ -24,11 +24,7 @@ from titiler.core.resources.enums import OptionalHeader
 from titiler.mosaic.errors import MOSAIC_STATUS_CODES
 from titiler.pgstac import __version__ as titiler_pgstac_version
 from titiler.pgstac.db import close_db_connection, connect_to_db
-from titiler.pgstac.dependencies import (
-    CollectionIdParams,
-    ItemPathParams,
-    SearchIdParams,
-)
+from titiler.pgstac.dependencies import CollectionIdParams, ItemIdParams, SearchIdParams
 from titiler.pgstac.extensions import searchInfoExtension
 from titiler.pgstac.factory import (
     MosaicTilerFactory,
@@ -159,7 +155,7 @@ app.include_router(
 # STAC Item Endpoints
 stac = MultiBaseTilerFactory(
     reader=PgSTACReader,
-    path_dependency=ItemPathParams,
+    path_dependency=ItemIdParams,
     optional_headers=optional_headers,
     router_prefix="/collections/{collection_id}/items/{item_id}",
     add_viewer=True,